requirements.txt
streamlit==1.31.0
openai==1.12.0
azure-cognitiveservices-speech==1.35.0
python-dotenv==1.0.0
pydantic==2.6.1